conftest alignment
