+++ tests/unit/tests_standard_lib/test_fields_tests/test_field_bank.py
@@ -1,20 +1,6 @@
 import pytest
-from unittest.mock import MagicMock, mock_open
+from unittest.mock import MagicMock
 from pytest_splunk_addon.standard_lib.fields_tests.field_bank import FieldBank
-
-
-@pytest.fixture()
-def open_mock(monkeypatch):
-    open_mock = mock_open()
-    monkeypatch.setattr("builtins.open", open_mock)
-    return open_mock
-
-
-@pytest.fixture()
-def json_load_mock(monkeypatch):
-    load_mock = MagicMock()
-    monkeypatch.setattr("json.load", load_mock)
-    return load_mock
 
 
 @pytest.fixture()
